#1124 - Generic render method for annotation editors
- move render method from BratAnnotationEditor to AnnotationEditorBase
- move getLayersToRender method from BratAnnotationEditor to AnnotationEditorBase
