+++ webanno-api-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/annotation/AnnotationEditorBase.java
@@ -17,15 +17,37 @@
  */
 package de.tudarmstadt.ukp.clarin.webanno.api.annotation;
 
+import static de.tudarmstadt.ukp.clarin.webanno.api.WebAnnoConst.CHAIN_TYPE;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+
+import org.apache.uima.jcas.JCas;
+import org.apache.wicket.Page;
 import org.apache.wicket.ajax.AjaxRequestTarget;
+import org.apache.wicket.core.request.handler.IPageRequestHandler;
+import org.apache.wicket.core.request.handler.IPartialPageRequestHandler;
+import org.apache.wicket.event.Broadcast;
 import org.apache.wicket.markup.html.panel.Panel;
 import org.apache.wicket.model.IModel;
+import org.apache.wicket.request.cycle.RequestCycle;
+import org.apache.wicket.spring.injection.annot.SpringBean;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.JCasProvider;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.action.AnnotationActionHandler;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.AnnotatorState;
+import de.tudarmstadt.ukp.clarin.webanno.api.annotation.rendering.PreRenderer;
+import de.tudarmstadt.ukp.clarin.webanno.api.annotation.rendering.event.RenderAnnotationsEvent;
+import de.tudarmstadt.ukp.clarin.webanno.api.annotation.rendering.model.VAnnotationMarker;
+import de.tudarmstadt.ukp.clarin.webanno.api.annotation.rendering.model.VDocument;
+import de.tudarmstadt.ukp.clarin.webanno.api.annotation.rendering.model.VMarker;
+import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationLayer;
+import de.tudarmstadt.ukp.clarin.webanno.model.Mode;
+import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence;
+import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token;
 
 public abstract class AnnotationEditorBase
     extends Panel
@@ -33,6 +55,9 @@ public abstract class AnnotationEditorBase
     private static final Logger LOG = LoggerFactory.getLogger(AnnotationEditorBase.class);
     
     private static final long serialVersionUID = 8637373389151630602L;
+
+    private @SpringBean PreRenderer preRenderer;
+    private @SpringBean AnnotationEditorExtensionRegistry extensionRegistry;
 
     private AnnotationActionHandler actionHandler;
     private JCasProvider jcasProvider;
@@ -100,6 +125,66 @@ public abstract class AnnotationEditorBase
      * JSON via {@link AjaxRequestTarget#appendJavaScript(CharSequence)}.
      */
     protected abstract void render(AjaxRequestTarget aTarget);
+
+    protected VDocument render(JCas aJCas)
+    {
+        VDocument vdoc = new VDocument();
+        preRenderer.render(vdoc, getModelObject(), aJCas, getLayersToRender());
+
+        // Fire render event into backend
+        extensionRegistry.fireRender(aJCas, getModelObject(), vdoc);
+
+        // Fire render event into UI
+        Page page = null;
+        Optional<IPageRequestHandler> handler = RequestCycle.get().find(IPageRequestHandler.class);
+        if (handler.isPresent()) {
+            page = (Page) handler.get().getPage();
+        };
+
+        if (page == null) {
+            page = getPage();
+        }
+        send(page, Broadcast.BREADTH,
+                new RenderAnnotationsEvent(
+                        RequestCycle.get().find(IPartialPageRequestHandler.class).get(), aJCas,
+                        getModelObject(), vdoc));
+
+        if (isHighlightEnabled()) {
+            AnnotatorState state = getModelObject();
+
+            // Disabling for 3.3.0 by default per #406
+            // FIXME: should be enabled by default and made optional per #606
+            // if (state.getFocusUnitIndex() > 0) {
+            // response.addMarker(new SentenceMarker(Marker.FOCUS, state.getFocusUnitIndex()));
+            // }
+
+            if (state.getSelection().getAnnotation().isSet()) {
+                vdoc.add(new VAnnotationMarker(
+                        VMarker.FOCUS, state.getSelection().getAnnotation()));
+            }
+        }
+
+        return vdoc;
+    }
+
+    private List<AnnotationLayer> getLayersToRender()
+    {
+        AnnotatorState state = getModelObject();
+        List<AnnotationLayer> layersToRender = new ArrayList<>();
+        for (AnnotationLayer layer : state.getAnnotationLayers()) {
+            boolean isSegmentationLayer = layer.getName().equals(Token.class.getName())
+                    || layer.getName().equals(Sentence.class.getName());
+            boolean isUnsupportedLayer = layer.getType().equals(CHAIN_TYPE)
+                    && (state.getMode().equals(Mode.AUTOMATION)
+                    || state.getMode().equals(Mode.CORRECTION)
+                    || state.getMode().equals(Mode.CURATION));
+
+            if (layer.isEnabled() && !isSegmentationLayer && !isUnsupportedLayer) {
+                layersToRender.add(layer);
+            }
+        }
+        return layersToRender;
+    }
     
     public void setHighlightEnabled(boolean aValue)
     {
